fix(Database): use Zone scheduler for object and list factories
Since Firebase is adding WebSocket listeners before the Angular Zone exists,
all callbacks when data is received are being executed outside of Angular's
zone and are not triggering change detection. This fix uses the existing
ZoneScheduler, which is already used in auth.onAuth, to cause the observables
created by af.list() and af.object() to be emitted inside the current Zone at
the time of calling .list() or .object().

See angular/angular#4603 for progress on making RxJS more zone-aware.

Fixes #637